worker/requirements.txt
celery
redis
pymongo
requests
beautifulsoup4
python-dotenv
groq
trafilatura
youtube-transcript-api
pypdf
lxml
